LPD-58181 Merge SearchResponseTranslator/DefaultSearchResponseTranslator/SearchResponseTranslatorImpl
