+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/search/engine/adapter/search/SearchSearchResponseAssemblerImpl.java
@@ -13,6 +13,7 @@ import com.liferay.portal.search.aggregation.AggregationResults;
 import com.liferay.portal.search.aggregation.pipeline.PipelineAggregation;
 import com.liferay.portal.search.aggregation.pipeline.PipelineAggregationResultTranslator;
 import com.liferay.portal.search.document.DocumentBuilderFactory;
+import com.liferay.portal.search.elasticsearch7.internal.SearchHitDocumentTranslator;
 import com.liferay.portal.search.elasticsearch7.internal.aggregation.AggregationResultTranslatorFactory;
 import com.liferay.portal.search.elasticsearch7.internal.aggregation.ElasticsearchAggregationResultTranslator;
 import com.liferay.portal.search.elasticsearch7.internal.aggregation.ElasticsearchAggregationResultsTranslator;
@@ -20,12 +21,16 @@ import com.liferay.portal.search.elasticsearch7.internal.aggregation.PipelineAgg
 import com.liferay.portal.search.elasticsearch7.internal.aggregation.pipeline.ElasticsearchPipelineAggregationResultTranslator;
 import com.liferay.portal.search.elasticsearch7.internal.hits.SearchHitsTranslator;
 import com.liferay.portal.search.elasticsearch7.internal.search.response.SearchResponseTranslator;
+import com.liferay.portal.search.elasticsearch7.internal.stats.StatsTranslator;
 import com.liferay.portal.search.engine.adapter.search.SearchSearchRequest;
 import com.liferay.portal.search.engine.adapter.search.SearchSearchResponse;
 import com.liferay.portal.search.geolocation.GeoBuilders;
+import com.liferay.portal.search.groupby.GroupByResponseFactory;
 import com.liferay.portal.search.highlight.HighlightFieldBuilderFactory;
 import com.liferay.portal.search.hits.SearchHitBuilderFactory;
 import com.liferay.portal.search.hits.SearchHitsBuilderFactory;
+import com.liferay.portal.search.legacy.stats.StatsRequestBuilderFactory;
+import com.liferay.portal.search.legacy.stats.StatsResultsTranslator;
 import com.liferay.portal.search.searcher.SearchTimeValue;
 
 import java.util.List;
@@ -39,6 +44,7 @@ import org.elasticsearch.search.SearchHits;
 import org.elasticsearch.search.aggregations.Aggregations;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
 
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -95,6 +101,14 @@ public class SearchSearchResponseAssemblerImpl
 
 		return new ElasticsearchPipelineAggregationResultTranslator(
 			elasticsearchAggregation, _aggregationResults);
+	}
+
+	@Activate
+	protected void activate() {
+		_searchResponseTranslator = new SearchResponseTranslator(
+			_groupByResponseFactory, _searchHitDocumentTranslator,
+			_statsRequestBuilderFactory, _statsResultsTranslator,
+			_statsTranslator);
 	}
 
 	protected void setCount(
@@ -197,15 +211,29 @@ public class SearchSearchResponseAssemblerImpl
 	private GeoBuilders _geoBuilders;
 
 	@Reference
+	private GroupByResponseFactory _groupByResponseFactory;
+
+	@Reference
 	private HighlightFieldBuilderFactory _highlightFieldBuilderFactory;
 
 	@Reference
 	private SearchHitBuilderFactory _searchHitBuilderFactory;
 
 	@Reference
-	private SearchHitsBuilderFactory _searchHitsBuilderFactory;
+	private SearchHitDocumentTranslator _searchHitDocumentTranslator;
 
 	@Reference
+	private SearchHitsBuilderFactory _searchHitsBuilderFactory;
+
 	private SearchResponseTranslator _searchResponseTranslator;
+
+	@Reference
+	private StatsRequestBuilderFactory _statsRequestBuilderFactory;
+
+	@Reference
+	private StatsResultsTranslator _statsResultsTranslator;
+
+	@Reference
+	private StatsTranslator _statsTranslator;
 
 }
