+++ modules/apps/portal-search-opensearch2/portal-search-opensearch2-impl/src/main/java/com/liferay/portal/search/opensearch2/internal/search/engine/adapter/search/SearchSearchResponseAssemblerImpl.java
@@ -16,16 +16,21 @@ import com.liferay.portal.search.document.DocumentBuilderFactory;
 import com.liferay.portal.search.engine.adapter.search.SearchSearchRequest;
 import com.liferay.portal.search.engine.adapter.search.SearchSearchResponse;
 import com.liferay.portal.search.geolocation.GeoBuilders;
+import com.liferay.portal.search.groupby.GroupByResponseFactory;
 import com.liferay.portal.search.highlight.HighlightFieldBuilderFactory;
 import com.liferay.portal.search.hits.SearchHitBuilderFactory;
 import com.liferay.portal.search.hits.SearchHitsBuilderFactory;
+import com.liferay.portal.search.legacy.stats.StatsRequestBuilderFactory;
+import com.liferay.portal.search.legacy.stats.StatsResultsTranslator;
 import com.liferay.portal.search.opensearch2.internal.aggregation.AggregationResultTranslatorFactory;
 import com.liferay.portal.search.opensearch2.internal.aggregation.OpenSearchAggregationResultTranslator;
 import com.liferay.portal.search.opensearch2.internal.aggregation.OpenSearchAggregationResultsTranslator;
 import com.liferay.portal.search.opensearch2.internal.aggregation.OpenSearchPipelineAggregationResultTranslator;
 import com.liferay.portal.search.opensearch2.internal.aggregation.PipelineAggregationResultTranslatorFactory;
 import com.liferay.portal.search.opensearch2.internal.hits.HitsMetadataTranslator;
+import com.liferay.portal.search.opensearch2.internal.legacy.hits.HitDocumentTranslator;
 import com.liferay.portal.search.opensearch2.internal.search.response.SearchResponseTranslator;
+import com.liferay.portal.search.opensearch2.internal.stats.StatsTranslator;
 import com.liferay.portal.search.opensearch2.internal.util.SetterUtil;
 import com.liferay.portal.search.searcher.SearchTimeValue;
 
@@ -40,6 +45,7 @@ import org.opensearch.client.opensearch.core.SearchResponse;
 import org.opensearch.client.opensearch.core.search.HitsMetadata;
 import org.opensearch.client.opensearch.core.search.TotalHits;
 
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -93,6 +99,14 @@ public class SearchSearchResponseAssemblerImpl
 
 		return new OpenSearchPipelineAggregationResultTranslator(
 			aggregate, _aggregationResults);
+	}
+
+	@Activate
+	protected void activate() {
+		_searchResponseTranslator = new SearchResponseTranslator(
+			_groupByResponseFactory, _hitDocumentTranslator,
+			_statsRequestBuilderFactory, _statsResultsTranslator,
+			_statsTranslator);
 	}
 
 	protected void setCount(
@@ -193,7 +207,13 @@ public class SearchSearchResponseAssemblerImpl
 	private GeoBuilders _geoBuilders;
 
 	@Reference
+	private GroupByResponseFactory _groupByResponseFactory;
+
+	@Reference
 	private HighlightFieldBuilderFactory _highlightFieldBuilderFactory;
+
+	@Reference
+	private HitDocumentTranslator _hitDocumentTranslator;
 
 	@Reference
 	private SearchHitBuilderFactory _searchHitBuilderFactory;
@@ -201,7 +221,15 @@ public class SearchSearchResponseAssemblerImpl
 	@Reference
 	private SearchHitsBuilderFactory _searchHitsBuilderFactory;
 
-	@Reference
 	private SearchResponseTranslator _searchResponseTranslator;
+
+	@Reference
+	private StatsRequestBuilderFactory _statsRequestBuilderFactory;
+
+	@Reference
+	private StatsResultsTranslator _statsResultsTranslator;
+
+	@Reference
+	private StatsTranslator _statsTranslator;
 
 }
